Renamed agent.core to agent.utils for consistency.
